The i2dashboard S4 class.

The i2dashboard S4 class provides main functionality of the package. Besides global properties such as the dashboard title, author and theme, an instance of the i2dashboard class also stores individual dashboard pages and the navigation menu, as well as all components that make up the content of individual pages. A new instance can be created using the i2dashboard function.

# S4 method for i2dashboard
initialize(.Object, ...)

# S4 method for i2dashboard
show(object)

i2dashboard(...)

Arguments

.Object
An object of class i2dashboard.

...
Named slots of the i2dashboard object.

object
An i2dashboard class object.

Value

An i2dashboard object.

Slots

title

The dashboards title

author

The author of the dashboard

interactive

If a shiny-based report should be created

theme

The theme of the dashboard (see the documentation of flexdashboard for available themes)

datadir

Path to the directory, where report data is stored.

file

The output filename (recommend that the suffix should be '.Rmd').

pages

A list of dashboard pages

sidebar

Content of the global sidebar

colormaps

A named list with color mappings.

source

Either a logical value describing whether the source code should be embeded through an item in the navigation bar or a link to a URL where the source code can be found online.

share

A vector with any number of the following services: “facebook”, “twitter”, “google-plus”, “linkedin”, and “pinterest”. You can also specify “menu” to provide a generic sharing drop-down menu that includes all of the services.

navbar

A list of links in the navigation bar (see the documentation of flexdashboard).